Move data family Rule into Claim class
